refactoring string functions
