+++ cx/ast/ast.go
@@ -3,7 +3,6 @@ package ast
 import (
 	"errors"
 	"fmt"
-	"github.com/skycoin/cx/cx"
 	"github.com/skycoin/cx/cx/constants"
 	"github.com/skycoin/cx/cx/globals"
 	"github.com/skycoin/cx/cx/helper"
@@ -1113,25 +1112,3 @@ func (cxprogram *CXProgram) PrintProgram() {
 	fmt.Println(cxprogram.ToString())
 }
 
-// ToString returns the abstract syntax tree of a CX program in a
-// string format.
-func (cxprogram *CXProgram) ToString() string {
-	var ast string
-	ast += "Program\n" //why is top line "Program" ???
-
-	var currentFunction *CXFunction
-	var currentPackage *CXPackage
-
-	currentPackage, err := cxprogram.GetCurrentPackage();
-
-	if err != nil {
-		panic("CXProgram.ToString(): error, currentPackage is nil")
-	}
-
-	currentFunction, _ = cxprogram.GetCurrentFunction();
-	currentPackage.CurrentFunction = currentFunction
-
-	cxcore.buildStrPackages(cxprogram, &ast) //what does this do?
-
-	return ast
-}
